Added Project Cosmic Coder My Buisness
Since GST was implemented, small businessmen are under an immense pressure to calculate their expenditure, tax to be paid or applied to their business transactions. So, we built an android application in order to facilitate the businessmen.
With a friendly User Interface, we have provided users with the feature to calculate the amount of tax paid, the total profit or loss in a time interval with daily savings and expenditure as well.
The algorithm behind the working consists of simple mathematical formulae used for calculating the tax to be paid or levied depending upon the input provided by the user.
Therefore, with just the click of a button, we shall have our customer's work done.
For the beginning, we have included only a few categories to the application. However, the application is planned and designed to work with several categories of inputs. If provided with financial support further, we would definitely try to achieve our goals.